Update ShareAcitivy.
Replace with TextView to build bitmap.
Fix some bugs.
