+++ app/src/main/java/com/sctdroid/app/textemoji/discovery/SearchableActivity.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+import android.view.View;
 
 import com.sctdroid.app.textemoji.R;
 import com.sctdroid.app.textemoji.data.source.GifRepository;
@@ -23,6 +24,12 @@ public class SearchableActivity extends AppCompatActivity {
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         toolbar.setTitle(R.string.title_activity_search_result);
+        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                finish();
+            }
+        });
 
 
         // Get the intent, verify the action and get the query
